+++ src/main/java/com/herejava/book/service/BookService.java
@@ -71,7 +71,7 @@ public class BookService {
 	}
 
 	//예약번호로 객체 1개 가져오는 service 메소드
-	public Book selectOneBook(int bookNo) {
+	public Book selectOneBook(long bookNo) {
 		Connection conn = JDBCTemplate.getConnection();
 		BookDao dao = new BookDao();
 		Book b = dao.selectOneBook(conn, bookNo);
